go-fuzz-build: compile in all fuzz functions
Prior to this change, go-fuzz-build only accommodated a single fuzz function.
If it wasn't named Fuzz, you had to identify it with -func.
This was inconvenient for packages with multiple fuzz functions.

This change teaches go-fuzz-build to handle multiple fuzz functions.
Instead of requiring a function name to be provided,
go-fuzz-build examines all package-scoped objects and detects fuzz functions.
(A fuzz function is any function with an appropriate name and signature.)

For libfuzzer, the only real change is that if you have a package with
a single fuzz function that's not called Fuzz, you can now call go-fuzz-build
without specifying its name.

For go-fuzz, the change is more interesting.

If there are multiple fuzz functions, all of them are made available for use
when you run go-fuzz. You can select which one to use with the newly
introduced go-fuzz -func flag.

For backwards compatibility, if you provide go-fuzz-build with a -func flag,
it records that preference in the metadata file, and go-fuzz respects it.

And as with libfuzzer, if there's only a single fuzz function, you no longer
have to specify it.

Currently, we only expose the ability to select a single fuzz func to use
for any given invocation of go-fuzz.
However, the implementation communicates the fuzz function to use on every
single fuzz function invocation. This is in preparation for possible future
work in which go-fuzz runs multiple fuzz functions concurrently.

Updates #230