go-fuzz: call roundUpCover less
cover==0 is the most common case by far.
In that case, we can skip the overhead of calling roundUpCover.

This optimization doesn't have a particularly significant impact on go-fuzz
as it stands, but if you start to increase the coverage array size
(e.g. to fuzz large codebases like compilers), calling roundUpCover
here quickly becomes one of the major bottlenecks.